requirements.txt
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.0.3
